Bugs fixed, firebase added. Admin dashboard and functionalities added, navbar modifird to have search but with dummy data for now

+++ src/app/api/admin/products/[id]/route.ts
@@ -1,35 +1,6 @@
+import { auth } from '@/auth';
 import Product from '@/models/Product.model';
 import { NextRequest, NextResponse } from 'next/server';
-
-export async function GET(
-  request: NextRequest,
-  { params }: { params: tParams },
-) {
-  try {
-    const { id } = await params;
-    if (!id) {
-      return NextResponse.json(
-        { message: 'Product ID is required' },
-        { status: 400 },
-      );
-    }
-
-    const product = await Product.findById(id).populate('brand');
-    if (!product) {
-      return NextResponse.json(
-        { message: 'Product not found' },
-        { status: 404 },
-      );
-    }
-    return NextResponse.json(product);
-  } catch (error) {
-    console.error('GET /products/[id] error:', error);
-    return NextResponse.json(
-      { message: 'Internal Server Error' },
-      { status: 500 },
-    );
-  }
-}
 
 export async function PATCH(
   request: NextRequest,
@@ -43,6 +14,15 @@ export async function PATCH(
         { status: 400 },
       );
     }
+    const session = await auth();
+    if (!session)
+      return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+    if (!session.user.isAdmin)
+      return NextResponse.json(
+        { message: 'Insufficient Permission' },
+        { status: 403 },
+      );
+
     const body = await request.json();
     const product = await Product.findByIdAndUpdate(id, body, {
       new: true,
@@ -75,6 +55,14 @@ export async function DELETE(
         { status: 400 },
       );
     }
+    const session = await auth();
+    if (!session)
+      return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+    if (!session.user.isAdmin)
+      return NextResponse.json(
+        { message: 'Insufficient Permission' },
+        { status: 403 },
+      );
 
     await Product.findByIdAndDelete(id);
     return NextResponse.json({ message: 'Deleted sucessfully' });
